test: move oversized inline test modules
